refactor: extract profiler-specific VMStructs members to ProfilerVMStructsExt
