+++ ddprof-lib/src/main/cpp/livenessTracker.cpp
@@ -12,6 +12,7 @@
 #include "context.h"
 #include "context_api.h"
 #include "hotspot/vmStructs.h"
+#include "profilerVmStructsExt.h"
 #include "incbin.h"
 #include "jniHelper.h"
 #include "livenessTracker.h"
